requirements.txt
# SmartLead - AI-Powered Lead Scoring System
# Core API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# ML Frameworks
xgboost==2.0.3
torch==2.1.1
joblib==1.3.2

# Feature Store
feast==0.35.0
redis==5.0.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# MLflow
mlflow==2.11.0
pydantic==2.5.2

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.19.0

# Resilience
tenacity==8.2.3
pybreaker==0.7.0

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
uuid==1.30
